+++ src/main/java/com/glop/cibl_orga_sport/repository/RolesDao.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RolesDao extends JpaRepository<Roles, Long> {
+public interface RolesDao extends JpaRepository<Roles, Integer> {
     boolean existsByDesignation(String designation);
 
     Optional<Roles> findByDesignation(String designation);
